Added console script to terminate an environment in cloud snitch.
Termination is different from removal. Termination simply marks
the end of an environment.  All relationships that are still
current are set to either utcnow in milliseconds or to a specified
timestamp.

This is intended to be used for environments that no longer
physically exist and for enviroments with names that have changed.