Fixes issue where last_sync field is rendered as ms timestamp.
Alters the type service to allow components to discover the
display type for a given model and property. Available display
types so far are: datetime, file, and text.

Added a cs-field component that accepts a model, a property, and
a value and renderes the value accordingly. Files are rendered
in <pre><code></code></pre> tags. Text is rendered as is with no
alterations. Datetime as passed through the localtime filter.